security/gssapiauth: move from pkg/ccl/gssapiccl
Move the GSSAPI authentication package out of the CCL tree into
pkg/security/gssapiauth. The CCL tree separation is no longer necessary
since all CockroachDB code is now under the CockroachDB Software
License.

The package is named gssapiauth (rather than gssapi) for consistency
with the sibling jwtauth, ldapauth, and oidcauth packages.

The blank import is left in pkg/ccl/ccl_init.go for now; relocating it
to a server-level init point is tracked in a TODO there.

Informs #164337

Release note: None
