+++ pkg/BUILD.bazel
@@ -1014,7 +1014,6 @@ GO_TARGETS = [
     "//pkg/ccl/changefeedccl/timers:timers_test",
     "//pkg/ccl/changefeedccl:changefeedccl",
     "//pkg/ccl/changefeedccl:changefeedccl_test",
-    "//pkg/ccl/gssapiccl:gssapiccl",
     "//pkg/ccl/multiregionccl/multiregionccltestutils:multiregionccltestutils",
     "//pkg/ccl/multiregionccl:multiregionccl",
     "//pkg/ccl/multiregionccl:multiregionccl_test",
@@ -1951,6 +1950,7 @@ GO_TARGETS = [
     "//pkg/security/dyncert:dyncert",
     "//pkg/security/dyncert:dyncert_test",
     "//pkg/security/fips:fips",
+    "//pkg/security/gssapiauth:gssapiauth",
     "//pkg/security/jwtauth:jwtauth",
     "//pkg/security/jwtauth:jwtauth_test",
     "//pkg/security/jwthelper:jwthelper",
