+++ pkg/ccl/BUILD.bazel
@@ -8,7 +8,6 @@ go_library(
     deps = [
         "//pkg/backup",
         "//pkg/ccl/changefeedccl",
-        "//pkg/ccl/gssapiccl",
         "//pkg/ccl/multiregionccl",
         "//pkg/ccl/multitenantccl",
         "//pkg/ccl/workloadccl",
@@ -17,5 +16,6 @@ go_library(
         "//pkg/crosscluster/logical",
         "//pkg/crosscluster/physical",
         "//pkg/crosscluster/producer",
+        "//pkg/security/gssapiauth",
     ],
 )
