+++ pkg/security/gssapiauth/empty.go
@@ -3,7 +3,7 @@
 // Use of this software is governed by the CockroachDB Software License
 // included in the /LICENSE file.
 
-package gssapiccl
+package gssapiauth
 
 // This file is here so on platforms where gssapi.go isn't built there's at
 // least one .go file to build.
